Add files via upload
1) Homework_2_Rstats: 

Description: Solutions to the following questions - 

Random variables, normal distributions, variance, covariance, and correlations. Remember to show your R code and results in their entirety in Rmarkdown code blocks (8 points)
Using R, create a vector of 10 points randomly chosen from the normal distribution with expected mean 10 and standard deviation 5. Calculate the mean and variance. Plot a histogram of the points. (1)
Repeat with vectors of 20, 50, and 100 points. What do you notice happens to the mean, variance and the histogram as you increase the N? (1)
Repeat with a vector of 150 points. Save this vector as a variable in your R session called A1.(0.5)
Create another vector of 150 points. Save this vector as a variable in your R session called A2. (0.5)
Calculate a 2x2 covariance matrix between A1 and A2. Calculate the correlation between A1 and A2. We did not do either of these in class, so you will need to look these up. Please also write in a comment or plain text outside the code block what you did to find the appropriate function(s), where you found the information, and know how to use it. (2)
Create a new vector, A3, the same size as A1 and A2, where each element is the average of that element in vectors A1 and A2. Check that for an index i, where i is between 1 and 150, A3[i] equals 0.5*(A1[i] + A2[i]) Find the variance of A3. How does it compare to the variance of A1 and A2? Please give a neuroimaging example of what this could represent. When (in a neuroimaging setting) may you want to average two normally distributed variables instead of testing each independently? List 2 advantages and one disadvantage that doing this may have with respect to your study design. (3)
Data analyses and assumptions when using the general linear regression model. (2 points) In the R_Intro in class assignment, we used the lm() function to test for the association between the bilaterally averaged volume of the hippocampus and several dependent variables. While we removed data points that were clearly “bad” outliers, however we failed to ensure the linear model was an appropriate option for us. Please describe what we could (should) have done. For extra (not graded) practice, you may want to read in the data again and use R code to test whether one of the assumptions we made in running the linear model was in fact true.


2) Stuti_C_NIIN_Stats_Final 

Description: This was from the final exam for the R portion of the class. Solutions to the following questions - 

1. Pick one voxel from one control subject, and a voxel from an AD subject (it can be at the same or different position). We’ll call the intensity values at these voxels V1_CN and V1_AD, respectively. Now pick a second voxel from each of the images; we’ll label these V2_CN and V2_AD, respectively.
a. The voxels chosen should NOT be the center voxel shown by default when you open fslview. Pick different ones!
2. Let’s pretend V1_CN represents the mean value for a particular GM region of interest for control subjects and V1_AD, the mean value for the same GM ROI. Let’s pretend the standard deviation (not variance) for the controls is given as the absolute difference between V2_CN and V1_CN, and the standard deviation for the AD patients is the absolute difference between V2_AD and V1_AD. 
a.	Assuming the means and standard deviations you just calculated are from a normal distribution for each of the clinical diagnostic groups, draw 100 samples from each group. What is the mean and standard deviation for each of these? Create a histogram for each group. What statistical test would you use to determine if the means are significantly different between the two groups?
b.	What do you notice as the sample size gets larger?
3.	Now we will randomly assign age and sex to each of the 200 data samples you have. 
a.	The controls are between ages 45 and 70 years and the patients are between 55 and 75 years old. Randomly assign an age to each of the samples from the appropriate uniform distribution (i.e., between 45 and 70 or 55 and 75, depending on the diagnosis). 
b.	You also know that 45% of the controls are female and 60% of the patients are female, the rest are male. How could you draw from another set of uniform distributions to randomly assign sex for each of the 200 individuals to ensure the proportions are approximately what you know about the sample? Randomly assign each data point as male or female.  If you didn’t draw from the uniform distributions, explain what you did and your reasoning. What percent of each group are female? 
*Remember, since you are randomly drawing numbers, you may not get the expected percentages exactly – that’s ok! – but it should be close.
4.	Now we test for group differences while controlling for age and sex. Set up a linear regression to test whether there is a case/control difference between 100 AD patients and 100 controls for the GM measure you’ve selected (i.e., randomly drawn!), after covarying for age and sex.
a.	Explain in your own words what two assumptions we went over in class are needed to make sure the linear regression model is an appropriate model to use. Give examples of what may violate those assumptions in this case. 
b.	Let’s pretend we test for case/control differences in 4 additional regions of interest (all independent from each other). If our significance threshold for a single test is p < 0.05, what would be our threshold now? Why do we need to make this adjustment, and what are the dangers of not making this adjustment?